refactor: file handler 훅으로 분리 (#145)
